refactor(parsing): simplify merchant helpers
